# attendance-app-webapi/docs/swagger.yaml
basePath: /api
definitions:
  AttendanceResponse:
    properties:
      created_at:
        description: When the record was created
        example: "2025-10-21T09:00:00Z"
        type: string
      id:
        description: The unique identifier for the attendance record
        example: 1
        type: integer
      latitude:
        description: Latitude of recorded location
        example: -7.5583648316326295
        type: number
      longitude:
        description: Longitude of recorded location
        example: 110.8577696892991
        type: number
      notes:
        description: Notes from validator
        example: Attendance verified
        type: string
      photoURL:
        description: URL of the photo taken during attendance
        example: https://storage.example.com/photos/checkin123.jpg
        type: string
      updated_at:
        description: When the record was last updated
        example: "2025-10-21T09:00:00Z"
        type: string
      user_id:
        description: The user ID who created the attendance record
        example: 1
        type: integer
      validationStatus:
        allOf:
        - $ref: '#/definitions/models.ValidationStatus'
        description: Status of the attendance record
        example: PRESENT
    type: object
  AttendanceValidationRequest:
    properties:
      notes:
        description: Optional notes from the validator
        example: Leave early without permit
        type: string
      validationStatus:
        allOf:
        - $ref: '#/definitions/models.ValidationStatus'
        description: Status of the validation (PRESENT, ABSENT, LATE)
        example: REJECTED
    required:
    - validationStatus
    type: object
...
  models.AttendanceStatus:
    enum:
    - ON_TIME
    - LATE
    type: string
    x-enum-varnames:
    - OnTime
    - Late
  models.AttendanceSwagger:
    properties:
      CheckInLatitude:
        type: number
      CheckInLongitude:
        type: number
      CheckInPhotoURL:
        type: string
      CheckInTime:
        type: string
      CheckOutLatitude:
        type: number
      CheckOutLongitude:
        type: number
      CheckOutPhotoURL:
        type: string
      CheckOutTime:
        type: string
      CreatedAt:
        type: string
      ID:
        type: integer
      LocationID:
        type: integer
      Notes:
        type: string
      Status:
        $ref: '#/definitions/models.AttendanceStatus'
      UpdatedAt:
        type: string
      UserID:
        type: integer
      ValidationStatus:
        $ref: '#/definitions/models.ValidationStatus'
      ValidatorID:
        type: integer
    type: object
  models.ErrorResponse:
    properties:
      error:
        example: Error message description
        type: string
    type: object
...
  models.ValidationStatus:
    enum:
    - PENDING
    - PRESENT
    - ABSENT
    - LEAVE
    - REJECTED
    - DIDNT_CHECKOUT
    type: string
    x-enum-varnames:
    - Pending
    - Present
    - Absent
    - Leave
    - Rejected
    - DidntCheckout
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API service for digital attendance system with role-based access control.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Digital Attendance API
  version: "1.0"
paths:
...
  /user/attendance/check-in:
    post:
      consumes:
      - multipart/form-data
      description: Record user's check-in with photo and location. Location must be
        within defined radius of office coordinates.
      parameters:
      - description: Check-in photo (JPG, JPEG, PNG, max 5MB)
        in: formData
        name: photo
        required: true
        type: file
      - description: Location latitude
        in: formData
        name: latitude
        required: true
        type: number
      - description: Location longitude
        in: formData
        name: longitude
        required: true
        type: number
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created attendance record
          schema:
            $ref: '#/definitions/AttendanceResponse'
        "400":
          description: Invalid request, location too far from office, or invalid photo
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized or invalid token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "429":
          description: Already checked in today
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Check-in attendance
      tags:
      - attendance
  /user/attendance/check-out:
    post:
      consumes:
      - multipart/form-data
      description: Record user's check-out with photo and location
      parameters:
      - description: Check-out photo (JPG, JPEG, PNG, max 5MB)
        in: formData
        name: photo
        required: true
        type: file
      - description: Location latitude
        in: formData
        name: latitude
        required: true
        type: number
      - description: Location longitude
        in: formData
        name: longitude
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AttendanceSwagger'
        "400":
          description: Invalid request, location too far, or invalid photo
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: No check-in record found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Check-out attendance
      tags:
      - attendance
  /user/attendance/my-records:
    get:
      description: Retrieve all attendance records for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AttendanceSwagger'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get personal attendance records
      tags:
      - attendance
  /user/attendance/subordinates:
    get:
      description: Retrieve attendance records for all subordinates (supervisor only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AttendanceSwagger'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Not a supervisor
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get subordinate attendance records
      tags:
      - attendance
  /user/attendance/update/{id}:
    put:
      consumes:
      - application/json
      description: Update attendance record for a subordinate (supervisor only)
      parameters:
      - description: Attendance ID
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: validation
        required: true
        schema:
          $ref: '#/definitions/AttendanceValidationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AttendanceSwagger'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Not the supervisor
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Record not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update subordinate attendance record
      tags:
      - attendance
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: For testing, you can use this admin test token (valid for 1 year):<br>Bearer
      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwidXNlcm5hbWUiOiJhZG1pbiIsInJvbGUiOiJhZG1pbiIsImV4cCI6MTc5MzMyODc5OX0.sKEg9JeUxDqHyD4vGwAb6pk9iLrebJfBpTJGPRnMSrY
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"